Redesign of the config variables.
Removed the config cache and created files for each app in which the
config variables are defined.
